#157: use deleteReply to consolidate tinker flow
