+++ source/buttons/tinker.js
@@ -1,9 +1,11 @@
 const { ActionRowBuilder, StringSelectMenuBuilder } = require('discord.js');
 const { ButtonWrapper } = require('../classes');
 const { getGearProperty } = require('../gear/_gearDictionary');
-const { getAdventure } = require('../orcustrators/adventureOrcustrator');
-const { SAFE_DELIMITER } = require('../constants');
+const { getAdventure, setAdventure } = require('../orcustrators/adventureOrcustrator');
+const { SKIP_INTERACTION_HANDLING } = require('../constants');
 const { trimForSelectOptionDescription, listifyEN } = require('../util/textUtil');
+const { transformGear } = require('../util/delverUtil');
+const { renderRoom } = require('../util/embedUtil');
 
 const mainId = "tinker";
 module.exports = new ButtonWrapper(mainId, 3000,
@@ -28,7 +30,7 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				options.push({
 					label: gear.name,
 					description: trimForSelectOptionDescription(`Possibilities: ${listifyEN(sidegrades, true)}`),
-					value: `${gear.name}${SAFE_DELIMITER}${index}`
+					value: index.toString()
 				})
 			}
 		})
@@ -37,16 +39,46 @@ module.exports = new ButtonWrapper(mainId, 3000,
 			return;
 		}
 
-		interaction.reply({
-			content: "You can use 1 room action to change a piece of gear to a different upgrade.",
-			components: [
-				new ActionRowBuilder().addComponents(
-					new StringSelectMenuBuilder().setCustomId("randomtinker")
-						.setPlaceholder("Pick a piece of gear to randomly tinker with...")
-						.setOptions(options)
-				)
-			],
-			ephemeral: true
-		});
+		interaction.deferReply({ ephemeral: true }).then(() => {
+			return interaction.editReply({
+				content: "You can use 1 room action to change a piece of gear to a different upgrade.",
+				components: [
+					new ActionRowBuilder().addComponents(
+						new StringSelectMenuBuilder().setCustomId(`${SKIP_INTERACTION_HANDLING}${interaction.id}`)
+							.setPlaceholder("Pick a piece of gear to randomly tinker with...")
+							.setOptions(options)
+					)
+				]
+			});
+		}).then(reply => {
+			const collector = reply.createMessageComponentCollector({ max: 1 });
+			collector.on("collect", collectedInteraction => {
+				const adventure = getAdventure(interaction.channelId);
+				if (!adventure.room.hasResource("roomAction")) {
+					interaction.reply({ content: "The workshop's supplies have been exhausted.", ephemeral: true });
+					return;
+				}
+
+				const delver = adventure.delvers.find(delver => delver.id === interaction.user.id);
+				const unparsedIndex = collectedInteraction.values[0];
+				const index = parseInt(unparsedIndex);
+				const gearName = delver.gear[index].name;
+				/** @type {string[]} */
+				const sidegrades = getGearProperty(gearName, "sidegrades");
+				const sidegradeName = sidegrades[adventure.generateRandomNumber(sidegrades.length, "general")];
+				transformGear(delver, index, gearName, sidegradeName);
+				adventure.room.history.Tinkerers.push(delver.name);
+				adventure.room.decrementResource("roomAction", 1);
+				interaction.channel.send(`**${interaction.member.displayName}**'s *${gearName}* has been tinkered to **${sidegradeName}**!`);
+				setAdventure(adventure);
+				interaction.channel.messages.fetch(adventure.messageIds.room).then(roomMessage => {
+					return roomMessage.edit(renderRoom(adventure, interaction.channel));
+				});
+			})
+
+			collector.on("end", () => {
+				interaction.deleteReply();
+			})
+		})
 	}
 );
